improve: profile-tab, fix/complete ratings

+++ ride-sharing-backend/src/main/java/hu/ridesharing/controller/DriveController.java
@@ -80,7 +80,7 @@ public class DriveController {
     }
 
     @GetMapping("/drive-count")
-    public int getDriveCountByUsername(@RequestParam String username) {
+    public long getDriveCountByUsername(@RequestParam String username) {
         return journeyService.getDriveCountByUsername(username);
     }
